Add tests folder
